+++ auto-completion/pythonidelib.py
@@ -1,0 +1,26 @@
+def FlushOutput():
+    pass
+
+def GetGlobalDict() -> object:
+    pass
+
+def GetShortcutInfo() -> dict:
+    pass
+
+def GetTraceLevel() -> int:
+    pass
+
+def Greet() -> str:
+    pass
+
+def RefCount(arg1: object) -> int:
+    pass
+
+def SetCompletion(arg1: list):
+    pass
+
+def SetSyntaxInfo(arg1: list, arg2: list):
+    pass
+
+def SetTraceLevel(arg1: int):
+    pass